Refact: Separated some codes
