+++ src/app/showcase/edit/panel/ShowcaseEditor.style.tsx
@@ -1,0 +1,18 @@
+import { SxProps } from '@mui/system'
+
+const IconStyle: SxProps = {
+  color: 'text.normal',
+  width: '1rem',
+  height: '1rem',
+}
+
+const ShowcaseImageStyle: SxProps = {
+  width: '18.5rem',
+  height: '12.5rem',
+  objectFit: 'cover',
+  borderRadius: '0.25rem',
+  border: '1px solid',
+  borderColor: 'line.alternative',
+}
+
+export { IconStyle, ShowcaseImageStyle }
